test(integration): move the workspace fixture into fixtures.terraform
The workspace staging fixture (and VERSIONS_TF) now live in fixtures/terraform.py
alongside the other Terraform fixtures; test_examples.py keeps only the
parametrize data and the CRUD test.
